fix: move consts
