+++ packages/blocks/src/page-block/edgeless/components/component-toolbar/change-text-menu.ts
@@ -17,6 +17,7 @@ import {
 } from '../../../../_common/icons/index.js';
 import type { CssVariableName } from '../../../../_common/theme/css-variables.js';
 import { countBy, maxBy } from '../../../../_common/utils/iterable.js';
+import { CanvasTextFont } from '../../../../surface-block/consts.js';
 import {
   Bound,
   normalizeShapeBound,
@@ -27,7 +28,6 @@ import {
 } from '../../../../surface-block/index.js';
 import type { SurfaceBlockComponent } from '../../../../surface-block/surface-block.js';
 import type { EdgelessSelectionSlots } from '../../edgeless-page-block.js';
-import { CanvasTextFont } from '../../utils/consts.js';
 import type { EdgelessAlignPanel } from '../panel/align-panel.js';
 import {
   type ColorEvent,
