+++ packages/blocks/src/page-block/edgeless/components/panel/font-family-panel.ts
@@ -1,7 +1,7 @@
 import { css, html, LitElement } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import { CanvasTextFont } from '../../utils/consts.js';
+import { CanvasTextFont } from '../../../../surface-block/consts.js';
 
 @customElement('edgeless-font-family-panel')
 export class EdgelessFontFamilyPanel extends LitElement {
